chore: create server-side helpers
